+++ addons/web/static/src/core/lazy_component.js
@@ -1,0 +1,25 @@
+import { Component, onWillStart, xml } from "@odoo/owl";
+import { loadBundle } from "@web/core/assets";
+import { registry } from "@web/core/registry";
+
+/**
+ * Utility component that loads an asset bundle before instanciating a component
+ */
+export class LazyComponent extends Component {
+    static template = xml`<t t-component="Component" t-props="componentProps"/>`;
+    static props = {
+        Component: String,
+        bundle: String,
+        props: { type: [Object, Function], optional: true },
+    };
+    setup() {
+        onWillStart(async () => {
+            await loadBundle(this.props.bundle);
+            this.Component = registry.category("lazy_components").get(this.props.Component);
+        });
+    }
+
+    get componentProps() {
+        return typeof this.props.props === "function" ? this.props.props() : this.props.props;
+    }
+}
