+++ odoo/addons/test_assetsbundle/static/tests/lazy_component.test.js
@@ -1,8 +1,7 @@
 import { expect, test } from "@odoo/hoot";
-import { queryOne } from "@odoo/hoot-dom";
-import { mountWithCleanup } from "@web/../tests/web_test_helpers";
-import { LazyComponent } from "@web/core/assets";
 import { Component, xml } from "@odoo/owl";
+import { mountWithCleanup } from "@web/../tests/web_test_helpers";
+import { LazyComponent } from "@web/core/lazy_component";
 
 test("LazyComponent loads the required bundle", async () => {
     class Test extends Component {
@@ -21,8 +20,7 @@ test("LazyComponent loads the required bundle", async () => {
     await mountWithCleanup(Test);
     expect.verifySteps(["Lazy test component created"]);
     expect(".o_lazy_test_component").toHaveText("Lazy Component!");
-    expect(window.getComputedStyle(queryOne(".o_lazy_test_component")).backgroundColor).toBe(
-        "rgb(165, 94, 117)"
-    );
+    expect(".o_lazy_test_component:only").toHaveStyle({
+        backgroundColor: "rgb(165, 94, 117)",
+    });
 });
-
